- Fixed Settings menu uninitialized variable. - Fixed cross-platform GDTR/IDTR pointers. - Fixed unused relx/rely variables. - Fixed currentticks warning. - Fixed various internal BIOS interrupts warnings. - Fixed LBA conversion warnings. - Fixed logging size warning. - Fixed MID channel loading warning. - Fixed unused MMU bus reading. - Fixed unused NMI handling on MMU. - Fixed invaddr detection on MMU. - Implemented cross-platform 64-bit hexdecimal sprintf(LONGLONGSPRINTX) define, like LONGLONGSPRINTF. - Fixed PSP not knowing SDL_SwapLE32. - Fixed zalloc return value warning.

+++ UniPCemu/interrupts/interrupt19.c
@@ -45,7 +45,7 @@ extern ThreadParams_p BIOSMenuThread; //BIOS pause menu thread!
 
 void POSTThread()
 {
-	reset = EMU_BIOSPOST(); //Execute POST, process emulator reset if needed!
+	reset = (byte)EMU_BIOSPOST(); //Execute POST, process emulator reset if needed!
 }
 
 void BIOS_int19()
